ImGui Docking and Viewports
Cleaning hardcopied ImGui classes and functions and making it transparent, so that every change in the ImGui submodule won't affect engine side. This way we do not have to worry about key mappings on initialization and more importantly about events which are now handled by ImGui module itself, communicating with GLFW, hence having a possibility to dock windows and have viewports.