Timesteps
Add timesteps and delta time to OnUpdate() method of Layer class. Timestep is a float which is being calculated in Application class as a time that passed since the last frame. Also, move all the stuff from Application to SandBox project, so the Application class remain as more of a abstraction.